requirements.txt
flask==3.1.1
werkzeug==3.1.3
openai==1.30.5
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.3.0
python-dotenv==1.0.1
cryptography==42.0.5
Flask-WTF==1.2.2
bleach==6.2.0
wtforms==3.2.1
weasyprint==62.2
openpyxl==3.1.2

markdown
flask-talisman
